Replaced Note with DeleteNote component in deleted notes components
this was done so that each note component should best reflect its current state. For example, on press of a normal note, opens the note screen displaying the note contents while from the pressing a note from the deleted screen prompts the user when they want to restore or permanently delete.